+++ lib/iceworm/wormutil/mnbrak.c
@@ -17,14 +17,14 @@ void mnbrak(float *ax, float *bx, float *cx, float *fa, float *fb, float *fc, fl
                 SHFT(dum,*ax,*bx,dum)
                 SHFT(dum,*fb,*fa,dum)
         }
-        *cx=(*bx)+GOLD*(*bx-*ax);
+        *cx=(float)((*bx)+GOLD*(*bx-*ax));
         *fc=(*func)(*cx);
         while (*fb > *fc) {
                 r=(*bx-*ax)*(*fb-*fc);
                 q=(*bx-*cx)*(*fb-*fa);
-                u=(*bx)-((*bx-*cx)*q-(*bx-*ax)*r)/
-                        (2.0*SIGN(MAX(fabs(q-r),TINY),q-r));
-                ulim=(*bx)+GLIMIT*(*cx-*bx);
+                u=(float) ((*bx)-((*bx-*cx)*q-(*bx-*ax)*r)/
+                        (2.0*SIGN(MAX(fabs(q-r),TINY),q-r)));
+                ulim=(float)((*bx)+GLIMIT*(*cx-*bx));
                 if ((*bx-u)*(u-*cx) > 0.0) {
                         fu=(*func)(u);
                         if (fu < *fc) {
@@ -38,19 +38,19 @@ void mnbrak(float *ax, float *bx, float *cx, float *fa, float *fb, float *fc, fl
                                 *fc=fu;
                                 return;
                         }
-                        u=(*cx)+GOLD*(*cx-*bx);
+                        u=(float)((*cx)+GOLD*(*cx-*bx));
                         fu=(*func)(u);
                 } else if ((*cx-u)*(u-ulim) > 0.0) {
                         fu=(*func)(u);
                         if (fu < *fc) {
-                                SHFT(*bx,*cx,u,*cx+GOLD*(*cx-*bx))
+                                SHFT(*bx,*cx,u,(float)(*cx+GOLD*(*cx-*bx)))
                                 SHFT(*fb,*fc,fu,(*func)(u))
                         }
                 } else if ((u-ulim)*(ulim-*cx) >= 0.0) {
                         u=ulim;
                         fu=(*func)(u);
                 } else {
-                        u=(*cx)+GOLD*(*cx-*bx);
+                        u=(float)((*cx)+GOLD*(*cx-*bx));
                         fu=(*func)(u);
                 }
                 SHFT(*ax,*bx,*cx,u)
